split RecordSinkFactory into two implementations
